codestyle&layout corrections
