+++ app/src/main/java/com/polygon/myxtap/myxtapapplication/PersonProfileActivity.java
@@ -1,0 +1,14 @@
+package com.polygon.myxtap.myxtapapplication;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class PersonProfileActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_person_profile);
+    }
+
+}
