+++ app/src/main/res/layout/activity_menu_main.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    tools:context=".MainMenuActivity">
+
+    <Button
+        android:id="@+id/buttonSendMail"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:text="@string/button_send_mail" />
+    <Button
+        android:id="@+id/buttonOpenProfile"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:layout_below="@id/buttonSendMail"
+        android:layout_marginTop="16dp"
+        android:text="@string/button_open_profile" />
+
+</RelativeLayout>
